// client/src/assets/globalStyleVariables.ts
export enum GlobalColorsEnum {
    Transparent = '#0000',
    Green = 'green',
    Red = 'red',
    Main = 'black',
    BrightGreen = '#42bf39',
    LightGreen = '#33cc33',
    BrightRed = '#db1247',
    DarkYellow = '#c9c92a',
    DarkGray = '#282828',
    MainText = '#FFFFFF',
    SecondaryText = '#dfdfdf',
    AccentPink = '#d442f5',
    AccentViolet = '#8d42f5',
    LightGray = '#222222',
    LightPurple = '#7056B3',
    LightYellow = '#c9c92a',
    SimilarWebBlue = '#2b3d52',
    DarkBlue = '#6350ff',
    Inactive = 'rgba(166,166,166,0.73)',
}

export enum FontsEnum {
    Title = 'Satoshi-Variable',
    Subtitle = 'Satoshi-Variable',
    Paragraph = 'Roboto',
}